requirements.txt
# GlucoCast Requirements
# Python 3.9+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration
PyYAML>=6.0

# XML parsing (for OhioT1DM)
lxml>=4.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Development
black>=23.0.0
isort>=5.12.0
pytest>=7.3.0
pytest-cov>=4.0.0
flake8>=6.0.0
mypy>=1.0.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0

# Experiment tracking (optional)
# mlflow>=2.5.0
# wandb>=0.15.0
